// src/generated/types.ts
export type AgentPubKey = HoloHash;

/**
 * @public
 */
export type AnyDhtHash = HoloHash;

/**
 * @public
 */
export type AnyLinkableHash = HoloHash;

/**
 * The unique identifier for a Cell.
 * Cells are uniquely determined by this pair - this pair is necessary
 * and sufficient to refer to a cell in a conductor
 * @public
 */
export type CellId = [DnaHash, AgentPubKey];

/**
 * Ways of specifying a clone cell in the context of an app.
 * @public
 */
export type CloneCellId =
  { type: "clone_id"; value: CloneId } | { type: "dna_hash"; value: DnaHash };

/**
 * Identifier of a clone cell, composed of the DNA's role name and the index
 * of the clone, starting at 0.
 *
 * Example: `profiles.0`
 * @public
 */
export type CloneId = string;

/**
 * Cloned cell that was created from a provisioned cell at runtime.
 * @public
 */
export type ClonedCell = {
  /**
   * The cell's identifying data
   */
  cell_id: CellId;
  /**
   * A conductor-local clone identifier
   */
  clone_id: CloneId;
  /**
   * The hash of the DNA that this cell was instantiated from
   */
  original_dna_hash: DnaHash;
  /**
   * The DNA modifiers that were used to instantiate this clone cell
   */
  dna_modifiers: DnaModifiers;
  /**
   * The name the cell was instantiated with
   */
  name: string;
  /**
   * Whether or not the cell is running
   */
  enabled: boolean;
};

/**
 * @public
 */
export type DhtOpHash = HoloHash;

/**
 * @public
 */
export type DnaHash = HoloHash;

/**
 * @public
 */
export type DnaHashB64 = string;

/**
 * Modifiers of this DNA - the network seed, properties and origin time - as
 * opposed to the actual DNA code. These modifiers are included in the DNA
 * hash computation.
 * @public
 */
export type DnaModifiers = {
  /**
   * The network seed of a DNA is included in the computation of the DNA hash.
   * The DNA hash in turn determines the network peers and the DHT, meaning
   * that only peers with the same DNA hash of a shared DNA participate in the
   * same network and co-create the DHT. To create a separate DHT for the DNA,
   * a unique network seed can be specified.
   */
  network_seed: NetworkSeed;
  /**
   * Any arbitrary application properties can be included in this object.
   */
  properties: Uint8Array;
};
// ...
export type HoloHash = Uint8Array;

/**
 * @public
 */
export type HoloHashed<C> = { content: C; hash: HoloHash };

/**
 * @public
 */
export type InlineHash = HoloHash;

/**
 * @public
 */
export type MembraneProof = Uint8Array;

/**
 * @public
 */
export type NetworkSeed = string;