// src/generated/api/app/types.ts
export type AppInfo = {
  /**
   * The unique identifier for an installed app in this conductor
   */
  installed_app_id: string;
  /**
   * Info about the cells installed in this app. Lists of cells are ordered
   * and contain first the provisioned cell, then enabled clone cells and
   * finally disabled clone cells.
   */
  cell_info: { [key in string]: Array<CellInfo> };
  /**
   * The app's current status, in an API-friendly format
   */
  status: AppStatus;
  /**
   * The app's agent pub key.
   */
  agent_pub_key: AgentPubKey;
  /**
   * The original AppManifest used to install the app, which can also be used to
   * install the app again under a new agent.
   */
  manifest: AppManifest;
  /**
   * The timestamp when this app was installed.
   */
  installed_at: Timestamp;
};

/**
 * Represents the available conductor functions to call over an app interface
 * and will result in a corresponding [`AppResponse`] message being sent back over the
 * interface connection.
 *
 * # Errors
 *
 * Returns an [`AppResponse::Error`] with a reason why the request failed.
 * @public
 */
export type AppRequest =
  | { type: "app_info" }
  | {
      type: "agent_info";
      value: {
        /**
         * Optionally limit the results to specific DNA hashes
         */
        dna_hashes?: Array<DnaHash> | null;
      };
    }
  | {
      type: "add_agent_info";
      value: {
        /**
         * The signed agent info, as produced by [`kitsune2_api::AgentInfoSigned::encode`].
         */
        agent_infos: Array<string>;
      };
    }
  | {
      type: "peer_meta_info";
      value: {
        /**
         * The Kitsune2 url of the agent to query, as a string.
         */
        url: string;
        /**
         * Optionally limit the results to specific DNA hashes.
         */
        dna_hashes?: Array<DnaHash> | null;
      };
    }
  | { type: "call_zome"; value: ZomeCallParamsSigned }
  | { type: "get_countersigning_session_state"; value: CellId }
  | { type: "abandon_countersigning_session"; value: CellId }
  | { type: "publish_countersigning_session"; value: CellId }
  | { type: "create_clone_cell"; value: CreateCloneCellPayload }
  | { type: "disable_clone_cell"; value: DisableCloneCellPayload }
  | { type: "enable_clone_cell"; value: EnableCloneCellPayload }
  | {
      type: "dump_op_timings";
      value: {
        /**
         * The DNA whose DHT arc to dump op timings for. This app must run a
         * cell of this DNA.
         */
        dna_hash: DnaHash;
        /**
         * Pagination cursor from a previous `DumpOpTimings`; only ops
         * ordered strictly after it are returned. `None` starts from the
         * beginning.
         */
        cursor?: OpTimingsCursor | null;
        /**
         * Maximum number of ops to return. Must be greater than zero.
         */
        limit?: number | null;
      };
    }
  | {
      type: "dump_network_metrics";
      value: {
        /**
         * If set, limits the metrics dumped to a single DNA hash space.
         */
        dna_hash?: DnaHash | null;
        /**
         * Whether to include a DHT summary.
         *
         * You need a dump from multiple nodes in order to make a comparison, so this is not
         * requested by default.
         */
        include_dht_summary?: boolean;
      };
    }
  | { type: "dump_network_stats" }
  | { type: "list_wasm_host_functions" }
  | { type: "provide_memproofs"; value: MemproofMap }
  | { type: "enable_app" }
  | {
      type: "send_direct_signal";
      value: {
        /**
         * The app network to send messages on.
         */
        dna_hash: DnaHash;
        /**
         * The agents to send the signal payload to.
         *
         * An empty payload is treated as an error.
         */
        agents: Array<AgentPubKey>;
        /**
         * The signal payload.
         *
         * Treated as opaque by Holochain, it is up to the application to decide how to serialize,
         * deserialize and process payloads.
         */
        signal: Array<number>;
      };
    };
// ...
export type CellInfo =
  | { type: "provisioned"; value: ProvisionedCell }
  | { type: "cloned"; value: ClonedCell }
  | { type: "stem"; value: StemCell };
// ...
export type CreateCloneCellPayload = {
  /**
   * The DNA's role name to clone
   */
  role_name: string;
  /**
   * Modifiers to set for the new cell.
   * At least one of the modifiers must be set to obtain a distinct hash for
   * the clone cell's DNA.
   */
  modifiers: DnaModifiersOpt<YamlProperties>;
  /**
   * Optionally set a proof of membership for the clone cell
   */
  membrane_proof?: MembraneProof | null;
  /**
   * Optionally a name for the DNA clone
   */
  name?: string | null;
};

/**
 * Parameters to specify the clone cell to be disabled.
 * @public
 */
export type DisableCloneCellPayload = {
  /**
   * The clone id or cell id of the clone cell
   */
  clone_cell_id: CloneCellId;
};

/**
 * @public
 */
export type EnableCloneCellPayload = DisableCloneCellPayload;

/**
 * Provisioned cell, a cell instantiated from a DNA on app installation.
 * @public
 */
export type ProvisionedCell = {
  /**
   * The cell's identifying data
   */
  cell_id: CellId;
  /**
   * The DNA modifiers that were used to instantiate the cell
   */
  dna_modifiers: DnaModifiers;
  /**
   * The name the cell was instantiated with
   */
  name: string;
};
// ...
export type StemCell = {
  /**
   * The hash of the DNA that this cell would be instantiated from
   */
  original_dna_hash: DnaHash;
  /**
   * The DNA modifiers that will be used when instantiating the cell
   */
  dna_modifiers: DnaModifiers;
  /**
   * An optional name to override the cell's bundle name when instantiating
   */
  name: string | null;
};
// ...
export type ZomeCallParamsSigned = {
  /**
   * Bytes of the serialized zome call payload that consists of all fields of the
   * [`ZomeCallParams`].
   */
  bytes: ExternIO;
  /**
   * Signature by the provenance of the call, signing the bytes of the zome call payload.
   */
  signature: Signature;
};